Partial implementation of stopline conversion
Stop lines were already parsed as signals in road.signal. However they were
ignored as they had the type 294.
In this commit, they were extracted by making them an object of the StopLine
class in lanelets.py and then saved as an additional attribute of the
Network() class.
In the export_commonroad_scenario method, they were then accessed in the same way as signals
and signs and then added to the lanelets using setters.

TODO:
1. The stoplanes are currently not assigned to their respective lanelet.
2. The stoplanes do not have the proper 2 points as currently the same position
point is used as the two points that consititute the start and end of the stopline