Parsed t_road_signals_signalReference element of opendrive.
Currently, our converter did not parse this particular element of the opendrive format. This element is necessary as in order to avoid duplication, the opendrive format does not duplicate traffic signals and lights and instead simply assigns references.
Now that we have parsed this element we can use it in our conversion to assign all the missing signals lights and stoplines.