Extracted signal references for all roads
All the references were stored in a list object in the parameter groups class.
A map was also created since the signal id stored in the signal reference is the  one generated by opendrive
however in our case we change the id and hence we need to translate these old ids to new ones.