Added signal data in signalReference element to lanelets
signal reference data was stored int the parametric lanes group.
A map was created to map the signal ids in the opendrive format to the new lanelet format
The map was used to then assign the predecessor of each lanelet with the signal id that was retreived and then converted to the new id from the signalReference element.