Added traffic lights and traffic signs to lanelet network
All assignments of traffic lights and signs were done on the basis of proximit compared to incoming lanelets in all the intersections available in a lanelet network
Minimum distance was compared from the center point of the lanelet

Traffic light directions have yet to be defined.
Stop lines have to be assigned and referenced.